+++ xpms2d/class/MainCanvas.h
@@ -43,7 +43,7 @@ private:
   void	drawHVPS(P2d_rec *record, struct recStats &stats, float hdrVer, int probeNum, PostScript *ps);
   void	drawFast2DC(P2d_rec *record, struct recStats &stats, float hdrVer, int probeNum, PostScript *ps);
 
-  void	drawSlice(PostScript *ps, int i, unsigned long slice);
+  void	drawSlice(PostScript *ps, int i, uint32_t slice);
   void	drawSlice(PostScript *ps, int i, unsigned long long slice);
   void	enchiladaLineItem(PostScript *ps, int i, int cnt, Particle *cp);
